# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/account/login": {
            "post": {
                "description": "登陆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "登陆",
                "parameters": [
                    {
                        "description": "登录信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/attendance_record/approve/query/{leader_staff_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "查询主管待审批的考勤记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "主管员工ID",
                        "name": "leader_staff_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/attendance_record/approve_accept/{attendId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "审批通过考勤记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "考勤记录ID",
                        "name": "attendId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/attendance_record/approve_reject/{attendId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "审批拒绝考勤记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "考勤记录ID",
                        "name": "attendId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/attendance_record/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "创建考勤记录",
                "parameters": [
                    {
                        "description": "考勤记录",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AttendanceRecordCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/attendance_record/delete/{attendance_id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "删除考勤记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "考勤记录ID",
                        "name": "attendance_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/attendance_record/edit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "更新考勤记录",
                "parameters": [
                    {
                        "description": "更新考勤记录信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AttendanceRecordEditDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AttendanceRecordEditDTO"
                        }
                    }
                }
            }
        },
        "/api/attendance_record/get_attend_record_is_pay/{staff_id}/{date}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "查询员工指定月份考勤记录是否已结算",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "日期(格式：YYYY-MM)",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/attendance_record/query/{staff_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "查询员工考勤记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/attendance_record/query_history/{staff_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "查询员工历史考勤记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息"
                },
                "status": {
                    "description": "状态码",
                    "type": "boolean"
                }
            }
        },
...
        "model.AttendanceRecordCreateDTO": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "leave_days": {
                    "type": "integer"
                },
                "overtime_days": {
                    "type": "integer"
                },
                "staff_id": {
                    "type": "string"
                },
                "staff_name": {
                    "type": "string"
                },
                "work_days": {
                    "type": "integer"
                }
            }
        },
        "model.AttendanceRecordEditDTO": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "leave_days": {
                    "type": "integer"
                },
                "overtime_days": {
                    "type": "integer"
                },
                "staff_id": {
                    "type": "string"
                },
                "staff_name": {
                    "type": "string"
                },
                "work_days": {
                    "type": "integer"
                }
            }
        },
...
        "model.LoginDTO": {
            "type": "object",
            "required": [
                "branch_id",
                "staff_id",
                "user_password"
            ],
            "properties": {
                "branch_id": {
                    "type": "string"
                },
                "staff_id": {
                    "type": "string"
                },
                "user_password": {
                    "type": "string"
                }
            }
        },
...
    }
}